requirements.txt
feedparser==6.0.11
openai==1.40.0
tweepy==4.14.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
